Day 9/requirements.txt
langchain==0.1.17
langchain-community==0.0.38
langchain-core==0.1.53
langchain-google-genai==0.0.9
tavily-python==0.3.2
sentence-transformers==2.7.0
chromadb==0.4.24
pypdf==3.17.4
python-dotenv==1.0.1
huggingface-hub==0.22.2
transformers==4.40.1
accelerate==0.30.1
protobuf==4.25.3
gradio==4.26.0
plotly==5.22.0
